requirements.txt
python-telegram-bot==13.14
requests==2.28.2
python-dotenv==1.0.0
geopy==2.3.0
yandex-maps==0.1.3 